Add files via upload
Distribution companies face greater challenges in organizing their fleets efficiently. So they solve the problem by using VRT. The vehicle routing problem is a combinatorial optimization problem. The optimal set of routes for a group of vehicles to traverse is used to deliver to a given set of customers. The VRP has several variants depending on tasks performed and restrictions, such as time windows, multiple vehicles, backhauls, simultaneous delivery, pickup, etc. This paper establishes a model for a large-scale vehicle routing problem with simultaneous pickup and delivery, and the objective is to minimize total transportation costs